+++ styles/blast.h
@@ -45,8 +45,8 @@ public:
       int M = 1000 - T/FADEOUT_MS;
       if (M > 0) {
 	// TODO: Get rid of float math.
-	float dist = fabs(b.location - led/(float)num_leds_);
-	int N = fabs(dist - T / (WAVE_MS * 1000.0)) * WAVE_SIZE;
+	float dist = fabsf(b.location - led/(float)num_leds_);
+	int N = fabsf(dist - T / (WAVE_MS * 1000.0)) * WAVE_SIZE;
 	if (N <= 32) {
 	  mix += blast_hump[N] * M / 1000;
 	}
